fuzzer: first working version
Initial commit in this repo was just a port of official LibAFL sample
for QEMU-based fuzzing. This change introduces an actual CI-capable
fuzzer for Xen. Main changes:

 - Updated to latest LibAFL version
 - Added command-line options that simplify generic checks
 - Added -t option to make fuzzing timed
 - Added -r option to replay found crashes
 - Fuzzer now returns non-zero error code if it found a crash
 - Added `cargo make` scripts to fetch and build Xen and XTF
 - Added documentation
 - Added TODOs

As most of the changes are very intertwined, there is no sense in
splitting them into separate commits. Thus, we have this huge one.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>